+++ international/en-GB.php
@@ -767,6 +767,7 @@ language::add_translations([
 	"label_interrupt" =>					'Interrupt Play Queue',
 	"label_notwhilestopped" =>				'Not supported while playback is stopped',
 	"config_brave_credentials" =>			"Brave Search API Credentials",
+	"config_audiobook_genres" =>			"Detect Spoken Word by Genre(s)"
 
 	// ----------------------------------------------------
 	// Proobaly Unused